+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * conditional branches
   * character set intersection branches
   * (options were already correct for any children of such branches)
+- Fixed `Expression#match` and `#=~` not working with a single argument
 
 ### [1.5.0] - 2019-05-14 - [Janosch Müller](mailto:janosch84@gmail.com)